[WIP] beginning of rewrite
