+++ lib/how_is/sources.rb
@@ -1,0 +1,7 @@
+require "how_is"
+
+module HowIs
+  module Sources
+    # Simply for creating a namespace.
+  end
+end
