+++ lib/how_is/sources/github.rb
@@ -1,0 +1,9 @@
+require "how_is/sources"
+
+module HowIs
+  module Sources
+    class Github
+      # ...
+    end
+  end
+end
